pyproject.toml
[project]
name = "barcode"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "click==8.2.1",
    "fastapi==0.116.1",
    "h11==0.16.0",
    "idna==3.10",
    "numpy==2.2.6",
    "opencv-python==4.12.0.88",
    "pillow==11.3.0",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings>=2.10.1",
    "python-multipart>=0.0.20",
    "pyzbar==0.1.9",
    "requests>=2.32.5",
    "sniffio==1.3.1",
    "starlette==0.47.3",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.1",
    "uvicorn==0.35.0",
]
